+++ lib/src/app/view/screens/template/scaffold_screens.dart
@@ -2,10 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:andrious/src/view.dart';
-
 import 'package:andrious/src/controller.dart';
-
+import 'package:andrious/src/view.dart';
 import 'package:flutter/gestures.dart' show DragStartBehavior;
 
 /// The 'framework' of a typical Material Screen.